requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
CTkMessagebox==2.5
customtkinter==5.2.1
darkdetect==0.8.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.6
packaging==23.2
pillow==10.2.0
pyperclip==1.8.2
pytankerkoenig==0.0.7
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tkintermapview==1.29
urllib3==2.1.0

setuptools~=68.2.0
numpy~=1.26.3
matplotlib~=3.8.2
pandas~=2.2.0